Update time complexity and space complexity.
