+++ codes/cpp/chapter_computational_complexity/space_complexity.cpp
@@ -1,5 +1,5 @@
 /*
- * File: space_complexity_types.cpp
+ * File: space_complexity.cpp
  * Created Time: 2022-11-25
  * Author: Krahets (krahets@163.com)
  */
