+++ codes/cpp/chapter_computational_complexity/time_complexity.cpp
@@ -1,5 +1,5 @@
 /*
- * File: time_complexity_types.cpp
+ * File: time_complexity.cpp
  * Created Time: 2022-11-25
  * Author: Krahets (krahets@163.com)
  */
