--- codes/go/chapter_computational_complexity/time_complexity_types_test.go
@@ -1,49 +1,0 @@
-// File: time_complexity_types_test.go
-// Created Time: 2022-12-13
-// Author: cathay (cathaycchen@gmail.com)
-
-package chapter_computational_complexity
-
-import (
-	"fmt"
-	"testing"
-)
-
-func TestRunCount(t *testing.T) {
-	// ======= Test Case =======
-	n := 8
-	fmt.Println("输入数据大小 n =", n)
-
-	// ====== Driver Code ======
-	count := constant(n)
-	fmt.Println("常数阶的计算操作数量 =", count)
-
-	count = linear(n)
-	fmt.Println("线性阶的计算操作数量 =", count)
-	count = arrayTraversal(make([]int, n))
-	fmt.Println("线性阶（遍历数组）的计算操作数量 =", count)
-
-	count = quadratic(n)
-	fmt.Println("平方阶的计算操作数量 =", count)
-	nums := make([]int, n)
-	for i := 0; i < n; i++ {
-		nums[i] = n - i  // [n,n-1,...,2,1]
-	}
-	count = bubbleSort(nums)
-	fmt.Println("平方阶（冒泡排序）的计算操作数量 =", count)
-
-	count = exponential(n)
-	fmt.Println("指数阶（循环实现）的计算操作数量 =", count)
-	count = expRecur(n)
-	fmt.Println("指数阶（递归实现）的计算操作数量 =", count)
-
-	count = logarithmic(float32(n))
-	fmt.Println("对数阶（循环实现）的计算操作数量 =", count)
-	count = logRecur(float32(n))
-	fmt.Println("对数阶（递归实现）的计算操作数量 =", count)
-	count = linearLogRecur(float32(n))
-	fmt.Println("线性对数阶（递归实现）的计算操作数量 =", count)
-
-	count = factorialRecur(n)
-	fmt.Println("阶乘阶（递归实现）的计算操作数量 =", count)
-}
