+++ codes/go/chapter_computational_complexity/worst_best_time_complexity.go
@@ -1,12 +1,15 @@
 // File: worst_best_time_complexity.go
 // Created Time: 2022-12-13
-// Author: cathay (cathaycchen@gmail.com)
+// Author: msk397 (machangxinq@gmail.com), cathay (cathaycchen@gmail.com)
 
 package chapter_computational_complexity
 
-import "math/rand"
+import (
+	"fmt"
+	"math/rand"
+)
 
-// randomNumbers 生成一个数组，元素为 { 1, 2, ..., n }，顺序被打乱
+/* 生成一个数组，元素为 { 1, 2, ..., n }，顺序被打乱 */
 func randomNumbers(n int) []int {
 	nums := make([]int, n)
 	// 生成数组 nums = { 1, 2, 3, ..., n }
@@ -20,7 +23,7 @@ func randomNumbers(n int) []int {
 	return nums
 }
 
-// findOne 查找数组 nums 中数字 1 所在索引
+/* 查找数组 nums 中数字 1 所在索引 */
 func findOne(nums []int) int {
 	for i := 0; i < len(nums); i++ {
 		if nums[i] == 1 {
@@ -28,4 +31,15 @@ func findOne(nums []int) int {
 		}
 	}
 	return -1
+}
+
+/* Driver Code */
+func main() {
+	for i := 0; i < 10; i++ {
+		n := 100
+		nums := randomNumbers(n)
+		index := findOne(nums)
+		fmt.Println("\n数组 [ 1, 2, ..., n ] 被打乱后 =", nums)
+		fmt.Println("数字 1 的索引为", index)
+	}
 }
