+++ codes/go/chapter_computational_complexity/worst_best_time_complexity_test.go
@@ -1,6 +1,6 @@
-// Copyright 2022 Cathay.  All rights reserved.
-// Use of this source code is governed by a MIT style
-// license that can be found in the LICENSE file.
+// File: worst_best_time_complexity.go
+// Created Time: 2022-12-13
+// Author: msk397 (machangxinq@gmail.com), cathay (cathaycchen@gmail.com)
 
 package chapter_computational_complexity
