+++ codes/java/chapter_computational_complexity/time_complexity.java
@@ -1,12 +1,12 @@
 /*
- * File: time_complexity_types.java
+ * File: time_complexity.java
  * Created Time: 2022-11-25
  * Author: Krahets (krahets@163.com)
  */
 
 package chapter_computational_complexity;
 
-public class time_complexity_types {
+public class time_complexity {
     /* 常数阶 */
     static int constant(int n) {
         int count = 0;
