+++ codes/python/chapter_computational_complexity/space_complexity.py
@@ -1,5 +1,5 @@
 '''
-File: space_complexity_types.py
+File: space_complexity.py
 Created Time: 2022-11-25
 Author: Krahets (krahets@163.com)
 '''
