+++ codes/python/chapter_computational_complexity/time_complexity.py
@@ -1,5 +1,5 @@
 '''
-File: time_complexity_types.py
+File: time_complexity.py
 Created Time: 2022-11-25
 Author: Krahets (krahets@163.com)
 '''
